Increasing pet movement speed
When the pet is moving from off screen, he will now be transported
closer to his destination.

Also, bug fix for fire button sticking around after a gate is cleared.

This introduces a new type of click lock.  If the camera is moving (due
to a partition change) all clicking is disabled.