Item box I
Adding the first part of the first iteration of item boxes -- the boxes
themselves work (when supplied with the data), as well as xml work for
said data.  But there is still no save system for the boxes.  Also, the
boxes don't actually grant items yet (just the UI works).