Bug fix: Degrad triggers turning on at wrong time
There was a bug where, during the tutorial, if you entered the first
gated room and then left it, the tutorial trigger would begin to attack.
Now fixed.

As an added bonus, I also made it so that the keyboard shortcuts for
moving on PC now obey the same rules as on the iPad.  Huzzah!