Changing spotlight instantiation
Prior to this change spotlights were always created in the center
anchor, but this was actually overly complicated and didn't work on
different aspect ratios other than 4:3.  This change now instantiates
them in the center for 3D objects, and within the anchor of a NGUI
object they may be highlighting.